demo10/requirements.txt
openai
fpdf2
streamlit
